# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/billing/checkout": {
      "post": {
        "operationId": "billingCheckout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Stripe-hosted checkout URL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutUrlDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed — malformed or out-of-bounds request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "Plan upgrade required — current plan does not include this feature.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "admin role required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found (or deliberately hidden to avoid an existence oracle).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "406": {
            "description": "No acceptable representation for the requested `Accept` header.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported request media type.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded — too many requests; retry after the `Retry-After` interval.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "default": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a Stripe Checkout session to subscribe to Pro or Enterprise",
        "tags": [
          "billing"
        ]
      }
    },
    "/billing/portal": {
      "post": {
        "operationId": "billingPortal",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Stripe Billing Portal URL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortalUrlDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed — malformed or out-of-bounds request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "Plan upgrade required — current plan does not include this feature.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "admin role required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found (or deliberately hidden to avoid an existence oracle).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "406": {
            "description": "No acceptable representation for the requested `Accept` header.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported request media type.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded — too many requests; retry after the `Retry-After` interval.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "default": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a Stripe Billing Portal session to manage the subscription",
        "tags": [
          "billing"
        ]
      }
    },
    "/billing/status": {
      "get": {
        "operationId": "billingStatus",
        "parameters": [],
        "responses": {
          "200": {
            "description": "billing and plan status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingStatusDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed — malformed or out-of-bounds request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "Plan upgrade required — current plan does not include this feature.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "admin role required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found (or deliberately hidden to avoid an existence oracle).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "406": {
            "description": "No acceptable representation for the requested `Accept` header.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported request media type.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded — too many requests; retry after the `Retry-After` interval.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "default": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get the current billing/plan status for the tenant",
        "tags": [
          "billing"
        ]
      }
    },
...
  },
  "info": {
    "title": "argus API",
    "description": "Crypto-blind delivery API for the argus E2EE messaging platform. Stores and forwards ciphertext only; never message content.",
    "version": "dev",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://api.argus.example.com"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CheckoutUrlDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "description": "Stripe-hosted checkout URL",
            "format": "uri",
            "pattern": "^[^\\u0000-\\u001f\\u007f]+$",
            "maxLength": 65536
          }
        },
        "required": [
          "url"
        ],
        "additionalProperties": false
      },
      "PortalUrlDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "description": "Stripe Billing Portal URL",
            "format": "uri",
            "pattern": "^[^\\u0000-\\u001f\\u007f]+$",
            "maxLength": 65536
          }
        },
        "required": [
          "url"
        ],
        "additionalProperties": false
      },
      "BillingStatusDto": {
        "type": "object",
        "properties": {
          "tier": {
            "type": "string",
            "enum": [
              "free",
              "pro",
              "enterprise"
            ]
          },
          "memberLimit": {
            "type": "integer",
            "nullable": true,
            "format": "int64",
            "maximum": 9007199254740991
          },
          "ssoEnabled": {
            "type": "boolean"
          },
          "memberCount": {
            "type": "integer",
            "format": "int64",
            "maximum": 9007199254740991
          },
          "subscriptionStatus": {
            "type": "string",
            "enum": [
              "active",
              "trialing",
              "past_due",
              "canceled",
              "incomplete"
            ],
            "nullable": true
          }
        },
        "required": [
          "tier",
          "memberLimit",
          "ssoEnabled",
          "memberCount",
          "subscriptionStatus"
        ],
        "additionalProperties": false
      },
...
      "ErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "statusCode",
          "message"
        ],
        "properties": {
          "statusCode": {
            "type": "integer",
            "format": "int64",
            "minimum": 100,
            "maximum": 599
          },
          "message": {
            "oneOf": [
              {
                "type": "string",
                "maxLength": 2048,
                "pattern": "^[^\\u0000-\\u001f\\u007f]+$"
              },
              {
                "type": "array",
                "maxItems": 100,
                "items": {
                  "type": "string",
                  "maxLength": 2048,
                  "pattern": "^[^\\u0000-\\u001f\\u007f]+$"
                }
              }
            ]
          },
          "error": {
            "type": "string",
            "maxLength": 256,
            "pattern": "^[^\\u0000-\\u001f\\u007f]+$"
          }
        }
      }
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}